tests/wpt/mozilla/tests/input-events/mousedown-edit-point-input-text.html

<!DOCTYPE html>
<title>Clicking in &lt;input type=text&gt; should move the edit point</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />

<input id="target" style="font: 50px/1 Ahem;" value="123456">

<script>
promise_test(async test => {
    let targetRect = target.getBoundingClientRect();
    let middleOfBlock = targetRect.top + (targetRect.height / 2);

    await new test_driver.Actions()
      .pointerMove(targetRect.left + 1, middleOfBlock)
      .pointerDown().pointerUp()
          .send();
    assert_equals(target.selectionStart, 0);
    assert_equals(target.selectionEnd, 0);

    await new test_driver.Actions()
      .pointerMove(targetRect.left + 50, middleOfBlock)
      .pointerDown().pointerUp()
          .send();
    assert_equals(target.selectionStart, 1);
    assert_equals(target.selectionEnd, 1);

    target.setSelectionRange(0, 0);

    // Click on the right half of a character should move the cursor to the next index.
    await new test_driver.Actions()
      .pointerMove(targetRect.left + 40, middleOfBlock)
      .pointerDown().pointerUp()
          .send();
    assert_equals(target.selectionStart, 1);
    assert_equals(target.selectionEnd, 1);

    target.setSelectionRange(0, 0);

    // Click on the right half of a character should move the cursor to the next index.
    await new test_driver.Actions()
      .pointerMove(targetRect.left + (50 * 5), middleOfBlock)
      .pointerDown().pointerUp()
          .send();
    assert_equals(target.selectionStart, 5);
    assert_equals(target.selectionEnd, 5);

    target.setSelectionRange(0, 0);

    // Clicking past the end of the text should move the edit point to the end.
    await new test_driver.Actions()
      .pointerMove(targetRect.left + (50 * 7), middleOfBlock)
      .pointerDown().pointerUp()
          .send();
    assert_equals(target.selectionStart, 6);
    assert_equals(target.selectionEnd, 6);
}, 'Mouse events move the edit point in <input type=text>');
</script>
